move kernels to kernels directory
